% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R, R/play.R
\name{authors}
\alias{authors}
\alias{is.authors}
\alias{summary.authors}
\alias{summary.play_igraph}
\title{Retrieve authors and number of plays they have for a corpus}
\usage{
authors(corpus)

\method{is}{authors}(x)

\method{summary}{authors}(object, ...)

\method{summary}{play_igraph}(object, ...)
}
\arguments{
\item{corpus}{\code{corpus} object or corpus name as character. Vector of corpus
names is not supported}

\item{x}{an R object}

\item{object}{An object of class \code{"authors"}.}

\item{...}{Other arguments to be passed to \code{\link{summary.default}}.}

\item{object}{An object of class \code{"play_igraph"}.}
}
\value{
\code{authors} object that inherits data.frame (and can be used
  as such).
}
\description{
\code{authors} returns \code{authors} object that inherits data.frame. For
each author in a corpus her/his respective number of plays is calculated.
}
\details{
There are more than one author for a play in some cases (e.g., in German
Drama Corpus). Coauthorship is not "weighted" for these plays: for each
author a play is counted as one play. Thus, a sum of all plays by al writers
in \code{authors} can be more than a number of plays in a corpus.
}
\section{Methods (by generic)}{
\itemize{
\item \code{is}: Tests that object is \code{authors}.

\item \code{summary}: Meaningful summary for \code{authors} object

\item \code{summary}: Meaningful summary for \code{play_igraph} object:
network properties, gender distribution
}}

\examples{
ru <- get_corpus("rus")
ru_authors <- authors(ru)
summary(ru_authors)
}
