% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{get_coocur_igraph}
\alias{get_coocur_igraph}
\alias{plot.coocur_igraph}
\alias{summary.coocur_igraph}
\title{Retrieve an igraph co-occurrence network for a play.}
\usage{
get_coocur_igraph(play = NULL, corpus = NULL)

\method{plot}{coocur_igraph}(
  x,
  layout = igraph::layout_with_kk,
  vertex.label = label_coocur_igraph(x),
  gender_colors = c(MALE = "#0073C2", FEMALE = "#EFC000", UNKNOWN = "#99979D"),
  vertex_size_metric = c("numOfWords", "numOfScenes", "numOfSpeechActs", "degree",
    "weightedDegree", "closeness", "betweenness", "eigenvector"),
  vertex_size_scale = c(5, 20),
  edge_size_scale = c(0.5, 4),
  vertex.label.color = "#03070f",
  vertex.label.family = "sans",
  vertex.label.font = 2L,
  vertex.frame.color = "white",
  ...
)

\method{summary}{coocur_igraph}(object, ...)
}
\arguments{
\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector (longer than 1) is not supported.}

\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{x}{A \code{coocur_igraph} object to plot.}

\item{layout}{Function, an algorithm used for graph layout. See
\link[igraph]{igraph.plotting}.}

\item{vertex.label}{A character vector of character names. By default,
function \code{\link{label_coocur_igraph}} is used to avoid overplotting on
large graphs.}

\item{gender_colors}{Named vector with 3 values with colors for
MALE, FEMALE and UNKNOWN respectively. Set \code{NULL} to use default igraph
colors. If you set parameter \code{vertex.color} (see
\link[igraph]{igraph.plotting}), \code{gender_colors} will be ignored.}

\item{vertex_size_metric}{Character value, one of \code{"numOfWords"},
\code{"numOfScenes"}, \code{"numOfSpeechActs"}, \code{"degree"},
\code{"weightedDegree"}, \code{"closeness"}, \code{"betweenness"},
\code{"eigenvector"}. You can specify vertex size by yourself using
parameter \code{"vertex.size"}(see \link[igraph]{igraph.plotting}), in this
case parameter \code{vertex_size_metric} is ignored.}

\item{vertex_size_scale}{Numeric vector with two values. The first number is
for mean size of node(vertex), the second one is for node size variance. If
you specify vertex size by yourself using parameter
\code{"vertex.size"}(see \link[igraph]{igraph.plotting}),
\code{vertex_size_scale} is ignored.}

\item{edge_size_scale}{Numeric vector with two values. The first number
defines average size of edges, the second number defines edges size variance.
If you specify edges size by yourself using parameter
\code{"edge.width"}(see \link[igraph]{igraph.plotting}),
\code{edge_size_scale} is ignored.}

\item{vertex.label.color}{See \link[igraph]{igraph.plotting}.}

\item{vertex.label.family}{See \link[igraph]{igraph.plotting}.}

\item{vertex.label.font}{See \link[igraph]{igraph.plotting}.}

\item{vertex.frame.color}{See \link[igraph]{igraph.plotting}.}

\item{...}{Other arguments to be passed to \link[igraph]{plot.igraph}}

\item{object}{An object of class \code{"coocur_igraph"}.}
}
\value{
\code{coocur_igraph} Object that inherits \code{igraph} and can be
treated as such.
}
\description{
Returns a play network, given corpus and play names. Play network is
constructed based on characters' co-occurrence matrix.
}
\section{Functions}{
\itemize{
\item \code{plot(coocur_igraph)}: Plot \code{coocur_igraph} using \code{plot.igraph}
with slightly modified defaults.

\item \code{summary(coocur_igraph)}: Meaningful summary for \code{"coocur_igraph"} object:
network properties, gender distribution

}}
\examples{
zhenitba_igraph <- get_coocur_igraph(play = "gogol-zhenitba", corpus = "rus")
igraph::diameter(zhenitba_igraph)
plot(zhenitba_igraph)
}
\seealso{
\code{\link{get_play_metadata}}
}
