% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracor_meta.R
\name{get_dracor_api_info}
\alias{get_dracor_api_info}
\title{Retrieve Dracor API info.}
\usage{
get_dracor_api_info()
}
\description{
\code{get_dracor_api_info} returns information about DraCor API: name of
the API, status, existdb version, and API version.
}
\details{
No parameters are expected.
}
\examples{
get_dracor_api_info()
}
