% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracor_meta.R
\name{get_dracor_meta}
\alias{get_dracor_meta}
\alias{summary.dracor_meta}
\alias{plot.dracor_meta}
\title{Retrieve information on available corpora}
\usage{
get_dracor_meta()

\method{summary}{dracor_meta}(object, ...)

\method{plot}{dracor_meta}(x, ...)
}
\arguments{
\item{object}{An object of class \code{"dracor_meta"}.}

\item{...}{Other arguments to be passed.}

\item{x}{A \code{dracor_meta} object.}
}
\value{
\code{dracor} object that inherits data.frame (and can be used
  as such).
}
\description{
\code{get_dracor_meta} returns a \code{dracor} object that inherits
data.frame (and can be used as such).
}
\section{Functions}{
\itemize{
\item \code{summary(dracor_meta)}: Meaningful summary for \code{dracor_meta} object.

\item \code{plot(dracor_meta)}: Plots how many plays are
available for each corpus.

}}
\examples{
corpora <- get_dracor_meta()
corpora
summary(corpora)
plot(corpora)
}
\seealso{
\code{\link{get_dracor}}
}
