% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{get_net_gexf}
\alias{get_net_gexf}
\title{Retrieve network for a play in GEXF.}
\usage{
get_net_gexf(play = NULL, corpus = NULL, parse = TRUE, ...)
}
\arguments{
\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector (longer than 1) is not supported.}

\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{parse}{Logical, if \code{TRUE} the result is parsed by
{\code{\link[xml2:read_xml]{xml2::read_xml()}}}, otherwise character value is
returned. Default value is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\value{
GEXF data.
}
\description{
The DraCor API lets you request a play network in GEXF (Graph Exchange XML
Format), given corpus and play names. GEXF is a format used in Gephi - an
open source software for network analysis and visualization.
}
\examples{
get_net_gexf(play = "gogol-zhenitba", corpus = "rus")
# If you want GEXF without parsing by xml2::read_xml():
get_net_gexf(play = "gogol-zhenitba", corpus = "rus", parse = FALSE)
}
\seealso{
\code{\link{get_play_metadata}}
}
