% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{get_net_igraph}
\alias{get_net_igraph}
\alias{plot.net_igraph}
\alias{summary.net_igraph}
\title{Retrieve an igraph network for a play.}
\usage{
get_net_igraph(play = NULL, corpus = NULL)

\method{plot}{net_igraph}(x, ...)

\method{summary}{net_igraph}(object, ...)
}
\arguments{
\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector (longer than 1) is not supported.}

\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{x}{A \code{net_igraph} object to plot.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}

\item{object}{An object of class \code{"net_igraph"}.}
}
\value{
\code{get_net_igraph} Object that inherits \code{igraph} and can be
treated as such.
}
\description{
Returns a play network, given corpus and play names. Play network is
constructed based on characters' cooccurrence matrix.
}
\section{Functions}{
\itemize{
\item \code{plot(net_igraph)}: Plot \code{net_igraph} using \code{plot.igraph}
with slightly modified defaults.

\item \code{summary(net_igraph)}: Meaningful summary for \code{"net_igraph"} object:
network properties, gender distribution

}}
\examples{
library(igraph)
zhenitba_igraph <- get_net_igraph(play = "gogol-zhenitba", corpus = "rus")
igraph::diameter(zhenitba_igraph)
plot(zhenitba_igraph)
}
\seealso{
\code{\link{get_play_metadata}}
}
