% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{get_play_metrics}
\alias{get_play_metrics}
\title{Retrieve network metrics for a play}
\usage{
get_play_metrics(corpus = NULL, play = NULL, ...)
}
\arguments{
\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector (longer than 1) is not supported.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\value{
List with network metrics for a specific play.
}
\description{
The DraCor API lets you request network metrics for a specific play, given
corpus and play names. Play network is constructed based on characters'
cooccurrence matrix.
}
\examples{
get_play_metrics(corpus = "rus", play = "gogol-zhenitba")
}
\seealso{
\code{\link{get_play_metadata}}, \code{\link{get_play_tei}},
  \code{\link{get_play_rdf}}, \code{\link{get_play_cast}},
  \code{\link{get_play_networkdata_csv}},
  \code{\link{get_play_networkdata_gexf}},
  \code{\link{get_play_spoken_text}},
  \code{\link{get_play_spoken_text_bych}},
  \code{\link{get_play_stage_directions}},
  \code{\link{get_play_stage_directions_with_sp}},
  \code{\link{play_igraph}}
}
