% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{get_play_networkdata_csv}
\alias{get_play_networkdata_csv}
\title{Retrieve edges list for a play.}
\usage{
get_play_networkdata_csv(corpus = NULL, play = NULL, ...)
}
\arguments{
\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor}}).}

\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_corpus}}). Character vector is not supported.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\description{
The DraCor API lets you request edges list for a play network, given corpus
and play names. Each row represents cooccurrences of two characters in a play
 - number of scenes where two characters appeared together. This edges list
can be used to construct a network for a play.
}
\examples{
get_play_networkdata_csv(corpus = "rus", play = "gogol-zhenitba")
}
\seealso{
\code{\link{get_play_metadata}}, \code{\link{get_play_metrics}},
  \code{\link{get_play_tei}}, \code{\link{get_play_rdf}},
  \code{\link{get_play_cast}},
  \code{\link{get_play_networkdata_gexf}},
  \code{\link{get_play_spoken_text}},
  \code{\link{get_play_spoken_text_bych}},
  \code{\link{get_play_stage_directions}},
  \code{\link{get_play_stage_directions_with_sp}},
  \code{\link{play_igraph}}
}
