% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{get_play_networkdata_gexf}
\alias{get_play_networkdata_gexf}
\title{Retrieve network for a play in GEXF.}
\usage{
get_play_networkdata_gexf(corpus = NULL, play = NULL, ...)
}
\arguments{
\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector is not supported.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\value{
GEXF data parsed by {\code{\link[xml2:read_xml]{xml2::read_xml()}}}.
}
\description{
The DraCor API lets you request a play network in GEXF (Graph Exchange XML
Format), given corpus and play names. GEXF is a format used in Gephi - an
open source software for network analysis and visualization.
}
\examples{
get_play_networkdata_gexf(corpus = "rus", play = "gogol-zhenitba")
# If you want GEXF without parsing by xml2::read_xml():
get_play_networkdata_gexf(corpus = "rus", play = "gogol-zhenitba", parse = FALSE)
}
\seealso{
\code{\link{get_play_metadata}}, \code{\link{get_play_metrics}},
  \code{\link{get_play_tei}}, \code{\link{get_play_rdf}},
  \code{\link{get_play_cast}},
  \code{\link{get_play_networkdata_csv}},
  \code{\link{get_play_spoken_text}},
  \code{\link{get_play_spoken_text_bych}},
  \code{\link{get_play_stage_directions}},
  \code{\link{get_play_stage_directions_with_sp}},
  \code{\link{play_igraph}}
}
