% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{get_play_rdf}
\alias{get_play_rdf}
\title{Retrieve an RDF for a play.}
\usage{
get_play_rdf(corpus = NULL, play = NULL, ...)
}
\arguments{
\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector is not supported.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\value{
RDF data parsed by {\code{\link[xml2:read_xml]{xml2::read_xml()}}}.
}
\description{
The DraCor API lets you request a RDF (Resource Description Framework) data
for a play, given corpus and play names. RDF for plays can be useful for
extraction data for a play from \url{Wikidata.org}.
}
\examples{
get_play_rdf(corpus = "rus", play = "gogol-zhenitba")
# If you want RDF without parsing by xml2::read_xml():
get_play_rdf(corpus = "rus", play = "gogol-zhenitba", parse = FALSE)
}
\seealso{
\code{\link{get_play_metadata}}, \code{\link{get_play_metrics}},
  \code{\link{get_play_tei}}, \code{\link{get_play_cast}},
  \code{\link{get_play_networkdata_csv}},
  \code{\link{get_play_networkdata_gexf}},
  \code{\link{get_play_spoken_text}},
  \code{\link{get_play_spoken_text_bych}},
  \code{\link{get_play_stage_directions}},
  \code{\link{get_play_stage_directions_with_sp}},
  \code{\link{play_igraph}}
}
