% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{get_play_spoken_text}
\alias{get_play_spoken_text}
\alias{get_play_spoken_text_bych}
\alias{get_play_stage_directions}
\alias{get_play_stage_directions_with_sp}
\title{Retrieve lines and stage directions for a play}
\usage{
get_play_spoken_text(corpus = NULL, play = NULL, gender = NULL, ...)

get_play_spoken_text_bych(corpus = NULL, play = NULL, ...)

get_play_stage_directions(corpus = NULL, play = NULL, ...)

get_play_stage_directions_with_sp(corpus = NULL, play = NULL, ...)
}
\arguments{
\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector is not supported.}

\item{gender}{Character, optional parameter to extract lines for characters
of specified gender: \code{"MALE"}, \code{"FEMALE"}, \code{"UNKNOWN"}.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\description{
The DraCor API lets you request lines and stage directions for a play, given
corpus and play names.
}
\section{Functions}{
\itemize{
\item \code{get_play_spoken_text_bych}: Retrieve lines grouped by characters in a
play, given corpus and play names.

\item \code{get_play_stage_directions}: Retrieve all stage directions of a play,
given corpus and play names.

\item \code{get_play_stage_directions_with_sp}: Retrieve all stage directions of a play
including speakers (if applicable), given corpus and play names.
}}

\examples{
get_play_spoken_text("rus", "gogol-zhenitba")
get_play_spoken_text("rus", "gogol-zhenitba", "FEMALE")
get_play_spoken_text_bych("rus", "gogol-zhenitba")
get_play_stage_directions("rus", "gogol-zhenitba")
get_play_stage_directions_with_sp("rus", "gogol-zhenitba")
}
\seealso{
\code{\link{get_play_metadata}}, \code{\link{get_play_metrics}},
  \code{\link{get_play_tei}}, \code{\link{get_play_rdf}},
  \code{\link{get_play_cast}},
  \code{\link{get_play_networkdata_csv}},
  \code{\link{play_igraph}}
}
