% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{get_play_tei}
\alias{get_play_tei}
\title{Retrieve a text for a play in TEI.}
\usage{
get_play_tei(corpus = NULL, play = NULL, ...)
}
\arguments{
\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor}}).}

\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_corpus}}). Character vector is not supported.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\value{
Text of a play parsed by
  {\code{\link[xml2:read_xml]{xml2::read_xml()}}}.
}
\description{
The DraCor API lets you request a text for a play in TEI format, given corpus
and play names. TEI is an XML vocabulary, which makes it easy to extract
structural information.
}
\examples{
get_play_tei(corpus = "rus", play = "gogol-zhenitba")
# If you want a text in TEI without parsing by xml2::read_xml():
get_play_tei(corpus = "rus", play = "gogol-zhenitba", parse = FALSE)
}
\seealso{
\code{\link{get_play_metadata}}, \code{\link{get_play_metrics}},
  \code{\link{get_play_rdf}}, \code{\link{get_play_cast}},
  \code{\link{get_play_networkdata_csv}},
  \code{\link{get_play_networkdata_gexf}},
  \code{\link{get_play_spoken_text}},
  \code{\link{get_play_spoken_text_bych}},
  \code{\link{get_play_stage_directions}},
  \code{\link{get_play_stage_directions_with_sp}},
  \code{\link{play_igraph}}
}
