% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{get_sparql}
\alias{get_sparql}
\title{Submit SPARQL queries to DraCor API.}
\usage{
get_sparql(sparql_query = NULL, ...)
}
\arguments{
\item{sparql_query}{Character, SPARQL query.
of specified gender: \code{"MALE"}, \code{"FEMALE"}, \code{"UNKNOWN"}.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\value{
SPARQL xml parsed by {\code{\link[xml2:read_xml]{xml2::read_xml()}}}.
}
\description{
The DraCor API lets you submit SPARQL queries.
}
\examples{
get_sparql("SELECT * WHERE {?s ?p ?o} LIMIT 10")
# If you want to avoid parsing by xml2::read_xml():
get_sparql("SELECT * WHERE {?s ?p ?o} LIMIT 10", parse = FALSE)
}
