% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{is.authors}
\alias{is.authors}
\title{Test an object to be an 'authors' object.}
\usage{
is.authors(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Test that object is an \code{\link{authors}}.
}
\seealso{
\code{\link{authors}}
}
