% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{is.coocur_igraph}
\alias{is.coocur_igraph}
\title{Test an object to be a 'coocur_igraph' object.}
\usage{
is.coocur_igraph(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Test that object is a \code{coocur_igraph}.
}
