% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{is.corpus}
\alias{is.corpus}
\title{Test an object to be a 'corpus' object.}
\usage{
is.corpus(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Test that object is a \code{corpus}.
}
\seealso{
\code{\link{get_corpus}}
}
