% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracor.R
\name{is.dracor}
\alias{is.dracor}
\title{Test an object to be a 'dracor' object.}
\usage{
is.dracor(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Test that object is a \code{dracor}.
}
\seealso{
\code{\link{get_dracor}}
}
