% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracor_meta.R
\name{is.dracor_meta}
\alias{is.dracor_meta}
\title{Test an object to be a 'dracor_meta' object.}
\usage{
is.dracor_meta(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Tests that object is a \code{dracor_meta}.
}
