% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{is.net_igraph}
\alias{is.net_igraph}
\title{Test an object to be a 'net_igraph' object.}
\usage{
is.net_igraph(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Test that object is a \code{net_igraph}.
}
