% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{is.play_igraph}
\alias{is.play_igraph}
\title{Test an object to be a 'play_igraph' object.}
\usage{
is.play_igraph(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Test that object is a \code{play_igraph}.
}
