% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{is.relations_igraph}
\alias{is.relations_igraph}
\title{Test an object to be a 'relations_igraph' object.}
\usage{
is.relations_igraph(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Tests that object is a \code{relations_igraph}.
}
