% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{label_net_igraph}
\alias{label_net_igraph}
\title{Extract labels for plotting 'net_igraph' object.}
\usage{
label_net_igraph(graph, max_graph_size = 30L, top_nodes = 3L)
}
\arguments{
\item{graph}{\code{net_igraph} object to plot.}

\item{max_graph_size}{Integer, maximum network size for plotting all labels}

\item{top_nodes}{Integer, number of labels to be plotted. Characters with the
highest number of words will be selected.}
}
\value{
Character vector of character names.
}
\description{
Extract labels for plotting \code{'net_igraph'} object. \code{label_net_igraph}
gives control of overplotting for labels (i.e. character names) by deleting
extra labels if there are too many of them. This function can be used to set
\code{vertex.label} parameter for \code{\link{plot.net_igraph}}.
}
\details{
\code{label_net_igraph} takes labels from a vertices data.frame column
\code{"name"}, checks that network size is more than \code{max_graph_size},
if it is true, returns names for top \code{top_nodes} and NA for the rest.
}
\examples{
library(igraph)
zhenitba_igraph <- get_net_igraph(play = "gogol-zhenitba", corpus = "rus")
label_net_igraph(zhenitba_igraph, max_graph_size = 10, top_nodes = 4)
}
\seealso{
\code{\link{get_net_igraph}}
}
