% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{dracor_api}
\alias{dracor_api}
\title{Send a GET request to DraCor API and parse the results}
\usage{
dracor_api(request, expected_type = c("application/json",
  "application/xml", "text/csv", "text/plain"), parse = TRUE,
  default_type = FALSE, split_text = TRUE, ...)
}
\arguments{
\item{request}{Character, valid GET request.}

\item{expected_type}{Character, MIME type: one of \code{"application/json"},
\code{"application/xml"}, \code{"text/csv"}, \code{"text/plain"}.}

\item{parse}{Logical, if \code{TRUE} (default value), then a response is parsed depending on
\code{expected_type}. See details below.}

\item{default_type}{Logical, if \code{TRUE}, default response data type is
returned. Therefore, a response is not parsed and \code{parse} is ignored.
Default value is \code{FALSE}.}

\item{split_text}{Logical, if \code{TRUE}, plain text lines are read as
different values in a vector instead of returning one character value.
Default value is \code{TRUE}.}

\item{...}{Other arguments passed to a parser function.}
}
\description{
Sending a GET request to DraCor API with a specified expected type and parse
results depending on selected expected type.
}
\details{
There are four different MIME types (aka internet media type) that can be
retrieved for DraCor API, the specific combination of possible MIME types
depends on API command. When \code{parse = TRUE} is used, the content is
parsed depending on selected MIME type in \code{expected_type}:
\describe{
  \item{\code{application/json}}{\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}}
  \item{\code{application/xml}}{\code{\link[xml2:read_xml]{xml2::read_xml()}}}
  \item{\code{text/csv}}{\code{\link[data.table:fread]{data.table::fread()}}}
  \item{\code{text/plain}}{No need for additional preprocessing}}
}
