% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{dracor_api_info}
\alias{dracor_api_info}
\title{Retrieve Dracor API info.}
\usage{
dracor_api_info()
}
\description{
\code{dracor_api_info} returns information about DraCor API: name of
the API, status, existdb version, and API version.
}
\details{
No parameters are expected.
}
\examples{
dracor_api_info()
}
