% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{get_coocur_igraph}
\alias{get_coocur_igraph}
\alias{plot.coocur_igraph}
\alias{summary.coocur_igraph}
\title{Retrieve an igraph co-occurrence network for a play.}
\usage{
get_coocur_igraph(play = NULL, corpus = NULL)

\method{plot}{coocur_igraph}(x, vertex.label = label_coocur_igraph(x), ...)

\method{summary}{coocur_igraph}(object, ...)
}
\arguments{
\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector (longer than 1) is not supported.}

\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{x}{A \code{coocur_igraph} object to plot.}

\item{vertex.label}{A character vector of character names.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}

\item{object}{An object of class \code{"coocur_igraph"}.}
}
\value{
\code{coocur_igraph} Object that inherits \code{igraph} and can be
treated as such.
}
\description{
Returns a play network, given corpus and play names. Play network is
constructed based on characters' co-occurrence matrix.
}
\section{Functions}{
\itemize{
\item \code{plot(coocur_igraph)}: Plot \code{coocur_igraph} using \code{plot.igraph}
with slightly modified defaults.

\item \code{summary(coocur_igraph)}: Meaningful summary for \code{"coocur_igraph"} object:
network properties, gender distribution

}}
\examples{
zhenitba_igraph <- get_coocur_igraph(play = "gogol-zhenitba", corpus = "rus")
igraph::diameter(zhenitba_igraph)
plot(zhenitba_igraph)
}
\seealso{
\code{\link{get_play_metadata}}
}
