% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{get_corpus}
\alias{get_corpus}
\alias{get_corpus_all}
\alias{summary.corpus}
\title{Retrieve metadata for all plays in a corpus.}
\usage{
get_corpus(
  corpus = NULL,
  URL = paste0("https://dracor.org/api/corpora/", corpus),
  full_metadata = TRUE
)

get_corpus_all(full_metadata = TRUE)

\method{summary}{corpus}(object, ...)
}
\arguments{
\item{corpus}{Name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor}}).}

\item{URL}{Request URL.}

\item{full_metadata}{Logical: if \code{TRUE} (default value), then additional metadata are retrieved.}

\item{object}{An object of class \code{"corpus"}.}

\item{...}{Other arguments to be passed to \code{\link{summary.default}}.}
}
\value{
\code{corpus} object that inherits data.frame (and can be used as such)
}
\description{
The DraCor API lets you request data for plays for a specific corpus.
\code{get_corpus} returns \code{dracor} object that inherits
data.frame (and can be used as such) but specified \code{\link{summary}}
method and \code{\link{authors}} function.
}
\details{
\code{get_corpus} returns a \code{corpus} object that inherits
data.frame (and can be used as such).

\code{corpus} constructs \code{corpus} object, \code{is.corpus} tests that
object is \code{corpus}, \code{summary.corpus} returns informative summary
for a corpus.

You need to provide a valid name for the corpus, e.g. \code{"rus"},
\code{"ger"} or \code{"shake"}. Use function \code{\link{get_dracor}}
to extract names for all available corpora.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Meaningful summary for \code{dracor} object.
}}

\examples{
\donttest{
ru <- get_corpus("rus")
head(ru)
summary(ru)
}
\donttest{
get_corpus_all()
}
}
\seealso{
\code{\link{get_dracor}}, \code{\link{authors}}
}
