% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracor.R
\name{get_dracor}
\alias{get_dracor}
\alias{summary.dracor}
\alias{plot.dracor}
\title{Retrieve information on available corpora.}
\usage{
get_dracor()

\method{summary}{dracor}(object, ...)

\method{plot}{dracor}(x, ...)
}
\arguments{
\item{object}{An object of class \code{"dracor"}.}

\item{...}{Other arguments to be passed.}

\item{x}{A \code{dracor} object.}
}
\value{
\code{dracor} object that inherits data.frame (and can be used
  as such).
}
\description{
\code{get_dracor} returns a \code{dracor} object that inherits
data.frame (and can be used as such).
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Meaningful summary for \code{dracor} object.

\item \code{plot}: Plots how many plays are
available for each corpus.
}}

\examples{
corpora <- get_dracor()
corpora
summary(corpora)
plot(corpora)
}
\seealso{
\code{\link{is.dracor}}, \code{\link{get_corpus}}
}
