% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{get_net_coocur_metrics}
\alias{get_net_coocur_metrics}
\title{Retrieve co-occurrence network metrics for a play}
\usage{
get_net_coocur_metrics(play = NULL, corpus = NULL, ...)
}
\arguments{
\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector (longer than 1) is not supported.}

\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\value{
List with network metrics for a specific play.
}
\description{
The DraCor API lets you request network metrics for a specific play, given
corpus and play names. Play network is constructed based on characters'
cooccurrence matrix.
}
\examples{
get_net_coocur_metrics(corpus = "rus", play = "gogol-zhenitba")
}
\seealso{
\code{\link{get_play_metadata}}
}
