% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{get_play_cast}
\alias{get_play_cast}
\title{Retrieve data for characters in a play.}
\usage{
get_play_cast(play = NULL, corpus = NULL, ...)
}
\arguments{
\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector (longer than 1) is not supported.}

\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\description{
The DraCor API lets you request miscellaneous information for characters in
a play, given corpus and play names: name, number and size of their lines,
gender, some network metrics etc.
}
\examples{
get_play_cast(play = "gogol-zhenitba", corpus = "rus")
}
\seealso{
\code{\link{get_play_metadata}}, \code{\link{get_play_metrics}},
  \code{\link{get_play_tei}}, \code{\link{get_play_rdf}},
  \code{\link{get_play_networkdata_csv}},
  \code{\link{get_play_networkdata_gexf}},
  \code{\link{get_play_spoken_text}},
  \code{\link{get_play_spoken_text_bych}},
  \code{\link{get_play_stage_directions}},
  \code{\link{get_play_stage_directions_with_sp}},
  \code{\link{play_igraph}}
}
