% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{get_text_chr_spoken}
\alias{get_text_chr_spoken}
\alias{get_text_chr_spoken_bych}
\alias{get_text_chr_stage}
\alias{get_text_chr_stage_with_sp}
\title{Retrieve lines and stage directions for a play}
\usage{
get_text_chr_spoken(
  play = NULL,
  corpus = NULL,
  gender = NULL,
  split_text = TRUE,
  ...
)

get_text_chr_spoken_bych(
  play = NULL,
  corpus = NULL,
  split_text = TRUE,
  dataframe = FALSE,
  ...
)

get_text_chr_stage(play = NULL, corpus = NULL, split_text = TRUE, ...)

get_text_chr_stage_with_sp(play = NULL, corpus = NULL, split_text = TRUE, ...)
}
\arguments{
\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector (longer than 1) is not supported.}

\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{gender}{Character, optional parameter to extract lines for characters
of specified gender: \code{"MALE"}, \code{"FEMALE"}, \code{"UNKNOWN"}.}

\item{split_text}{If \code{TRUE} returns text as a character vector of lines.
Otherwise, returns text as one character value. \code{TRUE} by default.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}

\item{dataframe}{If \code{TRUE} returns data frame with a row for every
character and text in a column \code{"text"}. \code{FALSE} by default.}
}
\description{
The DraCor API lets you request lines and stage directions for a play, given
corpus and play names.
}
\section{Functions}{
\itemize{
\item \code{get_text_chr_spoken_bych()}: Retrieve lines grouped by characters in a
play, given corpus and play names.

\item \code{get_text_chr_stage()}: Retrieve all stage directions of a play,
given corpus and play names.

\item \code{get_text_chr_stage_with_sp()}: Retrieve all stage directions of a play
including speakers (if applicable), given corpus and play names.

}}
\examples{
get_text_chr_spoken(play = "gogol-zhenitba", corpus = "rus")
get_text_chr_spoken(play = "gogol-zhenitba", corpus = "rus", "FEMALE")
get_text_chr_spoken(play = "gogol-zhenitba", corpus = "rus", "FEMALE", split = FALSE)
get_text_chr_spoken_bych(play = "gogol-zhenitba", corpus = "rus")
get_text_chr_stage(play = "gogol-zhenitba", corpus = "rus")
get_text_chr_stage_with_sp(play = "gogol-zhenitba", corpus = "rus")
}
\seealso{
\code{\link{get_play_metadata}}
}
