% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play_igraph}
\alias{play_igraph}
\alias{is.play_igraph}
\alias{plot.play_igraph}
\title{Retrieve an igraph network for a play.}
\usage{
play_igraph(corpus = NULL, play = NULL)

\method{is}{play_igraph}(x)

\method{plot}{play_igraph}(x, ...)
}
\arguments{
\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor}}).}

\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_corpus}}). Character vector is not supported.}

\item{x}{An R object.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}
}
\value{
\code{play_igraph} Object that inherits \code{igraph} and can be
treated as such.
}
\description{
Returns a play nertwork, given corpus and play names. Play network is
constructed based on characters' cooccurrence matrix.
}
\section{Methods (by generic)}{
\itemize{
\item \code{is}: Tests that object is \code{play_igraph}.

\item \code{plot}: Plot \code{play_igraph} using \code{play_igraph}
with slightly modified defaults.
}}

\examples{
library(igraph)
zhenitba_igraph <- play_igraph("rus", "gogol-zhenitba")
igraph::diameter(zhenitba_igraph)
plot(zhenitba_igraph)
}
\seealso{
\code{\link{get_play_metadata}}, \code{\link{get_play_metrics}},
  \code{\link{get_play_tei}}, \code{\link{get_play_rdf}},
  \code{\link{get_play_cast}},
  \code{\link{get_play_networkdata_csv}},
  \code{\link{get_play_networkdata_gexf}},
  \code{\link{get_play_spoken_text}},
  \code{\link{get_play_spoken_text_bych}},
  \code{\link{get_play_stage_directions}},
  \code{\link{get_play_stage_directions_with_sp}},
}
