% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play_igraph}
\alias{play_igraph}
\alias{plot.play_igraph}
\alias{summary.play_igraph}
\title{Retrieve an igraph network for a play.}
\usage{
play_igraph(play = NULL, corpus = NULL)

\method{plot}{play_igraph}(x, ...)

\method{summary}{play_igraph}(object, ...)
}
\arguments{
\item{play}{Character, name of the play (you can find all play names in
\code{playName} column within an object returned by
\code{\link{get_dracor}}). Character vector (longer than 1) is not supported.}

\item{corpus}{Character, name of the corpus (you can find all corpus names in
\code{name} column within an object returned by \code{\link{get_dracor_meta}}).}

\item{x}{A \code{play_igraph} object to plot.}

\item{...}{Additional arguments passed to \code{\link{dracor_api}}.}

\item{object}{An object of class \code{"play_igraph"}.}
}
\value{
\code{play_igraph} Object that inherits \code{igraph} and can be
treated as such.
}
\description{
Returns a play nertwork, given corpus and play names. Play network is
constructed based on characters' cooccurrence matrix.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot(play_igraph)}: Plot \code{play_igraph} using \code{play_igraph}
with slightly modified defaults.

\item \code{summary(play_igraph)}: Meaningful summary for \code{"play_igraph"} object:
network properties, gender distribution

}}
\examples{
library(igraph)
zhenitba_igraph <- play_igraph(play = "gogol-zhenitba", corpus = "rus")
igraph::diameter(zhenitba_igraph)
plot(zhenitba_igraph)
}
\seealso{
\code{\link{get_play_metadata}}, \code{\link{get_play_metrics}},
  \code{\link{get_play_tei}}, \code{\link{get_play_rdf}},
  \code{\link{get_play_cast}},
  \code{\link{get_play_networkdata_csv}},
  \code{\link{get_play_networkdata_gexf}},
  \code{\link{get_play_spoken_text}},
  \code{\link{get_play_spoken_text_bych}},
  \code{\link{get_play_stage_directions}},
  \code{\link{get_play_stage_directions_with_sp}},
  \code{\link{is.play_igraph}}
}
