% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{plot.authors}
\alias{plot.authors}
\title{Plot number of plays by authors of the corpus}
\usage{
\method{plot}{authors}(x, only_surnames = FALSE, top_n = nrow(x),
  top_minplays = 1, top_ratio = 1, ...)
}
\arguments{
\item{x}{\code{authors} object to plot.}

\item{only_surnames}{Logical, if \code{TRUE}, then only authors' surnames
are shown, defaults to \code{FALSE}.}

\item{top_n}{Integer, number of top authors to show on plot.}

\item{top_minplays}{Integer, minimum number of plays for an author to be
plotted.}

\item{top_ratio}{Numeric, from 0 to 1, maximum ratio of top authors to be
plotted. For example, with \code{top_ratio = 0.1} maximum 10\% of top
authors will be plotted.}
}
\description{
\code{plot.authors} plots numbers of plays for every author or only for
top authors.
}
\details{
There are some really long names, in case you want to plot authors with their
surnames only, use \code{only_surnames = TRUE}.
For large corpora, e.g., German Drama Corpus or Russian Drama Corpus,
there is a problem of overplotting. You can plot only top authors by
parameters \code{top_n}, \code{top_minplays}, or \code{top_ratio}
adjustment.
}
